internal/layer2: rename, again, from internal/arpndp. #183
After much bikeshedding, I've decided that this set of protocol
semantics and behaviors shall be called "layer2" throughout code,
config and documentation.

For people unfamiliar with the internals of networking, the name is
irrelevant, aside from the fact that it is short (for nav sidebars on
the website) and accurately describes the implementation.

For people with networking familiarity, "layer2" triggers all the right
associations: ethernet, ARP, NDP, VRRP, broadcasts, doesn't traverse
routers... All correct and useful associations to make in order to
quickly grasp what this balancing mode does.

As an additional benefit, the name is protocol-agnostic, meaning we
can in future support hypothetical dual-stacked Kubernetes clusters
with no change to configuration. Additionally, we can choose to
use additional L2 protocols in the implementation as needs arise
(e.g. VRRP to enable leader election without depending on k8s being
up).