Listen on all structurally valid interfaces, regardless of addresses.
There is a valid use case for having interfaces that are "services-only",
i.e. are marked up, but don't have global unicast addresses of their own.
Such addresses can still participate in layer2 mode advertisements, and
it can significantly save on the number of IPs in use on the network.

This change therefore listens on all interfaces that are up and support
the features required by ARP/NDP (e.g. broadcast, multicast, existence
of a link-local address...).

Fixes #226