tasks.py: Specify release verion in manifests
It seems that commit 416e01d79092d795a14c7db69bdce6917c3a9e01 removed
helm support but helm was also used to generate manifests/metallb.yaml.
Therefore, the manifests are not updated with the new release version
for the container images.

This patch fixes that by just adding a replace in the release script.