website: Reflect the move to memberlist
While reading about metallb via the web site, I noticed some content
that was out of date with the current implementation.  The layer 2
mode page still discussed the use of the Kubernetes API for leader
election.  I've updated the text to reflect that memberlist is now
used for cluster membership in metallb.