ci: Split lint out into its own job
gometalinter was previously run in the test job which runs "go test".
There should be a slight runtime improvement by moving this into its
own job, but the main reason to split it out is so it shows up as its
own status line item when results are reported back to github.  When
something fails, you will be able to more quickly determine which area
the failure is in (linting vs the tests).