Add exec permission in docker image
metallb.yaml runs controller as non-root user but in docker image,
controller binary is owned by root. If the environment this is built
happens to use a umask that masks execute permission for others, then
the user won't be able to execute.

This change makes sure that controller and speaker binaries can be
exeucted by non-root user in the docker image.

fixes #643

Ammended-By: Rodrigo Campos <rodrigo@kinvolk.io>
 Mention umask in the comment, change the speaker Dockerfile too,
extend the commit message.