layer2: Always ignore not ready nodes
A node should be ignored if it is not present in the map or if it is
present and not ready (i.e. present but with a false value).

This issue never triggers in practice today, though, because the
speakers map (created by UsableSpeakers() speakerlist method) always
contains entries with true values.

Signed-off-by: Rodrigo Campos <rodrigo@kinvolk.io>