Remove unused Service Update()
The Update() to the Service object is never triggered because the code
does not modify the object between the deep copy and when Update() is
called.

Previously, commit cdd44aa9 ("Initial commit. Somewhat working
controller, half-written BGP speaker") modified the Spec and Annotations
field of the Service. However, since commit 3e10f6ea ("Clean up processing
to exclusively use status.loadBalancer.ingress.ip"), that is no longer
the case, as the code exclusively only touches the Status field.

Therefore, this commit removes this dead code and revokes the update
privileges from the ClusterRole as well.

Signed-off-by: Chris Tarazi <chris@isovalent.com>