Introduce e2e test framework
This commit re-introduces system tests that were previously
removed.

These tests can be invoked via the `inv e2etest` command
and must be run against a configured "dev-env" development
environment.

One test has been added for both the "layer2" protocol and
the "bgp" protocol". The sample test deploys a service and
checks connectivity with the service. This is similar to what
is described in the dev-env README.md.

Further tests will be added in subsequent commits.

Signed-off-by: Mark Gray <mark.d.gray@redhat.com>