Adjust the l2 controller unit tests to the new logic
The l2 announcement logic changed to be consistent for ip, and different
if the traffic policy is local or cluster. Here we adjust the tests to
the new behaviour, changing the traffic policy to local which is more
similar to the old behaviour. Some tests change the controller
announcing because the hash changed.

Signed-off-by: Federico Paolinelli <fpaoline@redhat.com>