E2E: create the external containers only once (#1122)
* E2E: create the external containers only once

Instead of creating and destroying the containers after each run, we
create them only once and we keep them for all the tests.
This has several benefits:
- Reduce the test run time since we don't have to pay the container
  creation each time
- The test suite will clean the container even when we block the test
  execution
- This is likely to reduce / remove the flake that happens at container
  start time

Signed-off-by: Federico Paolinelli <fpaoline@redhat.com>

* E2E: Inject the container ips from outside

Now that we are starting the containers in the setup of the suite, we
are stealing two more ips from the available ip range. This collides
with L2 tests, which are using the free ip range that is passed from
outside. By fixing the ips used by the containers, we change the
behavior to be more deterministic and we avoid having overlapping ips
between the containers and the l2 services.

Signed-off-by: Federico Paolinelli <fpaoline@redhat.com>