FRR mode: fail in unsupported scenarios (#1150)
* Mark the config as non valid if routerid is specified in FRR mode

Currently it's not possible to set different routerids per different
peers with FRR. The rendered config will contain a single routerid (the
speaker's one) per node, giving a result different from what the user
expects (a specific routerid per peer). For this reason, we remove the
configuration logic and we reject a configuration containing the
RouterID field when being applied in FRR mode.

Signed-off-by: Federico Paolinelli <fpaoline@redhat.com>

* Reject the configuration if multiple ASN are configured in frr mode

FRR does not allow to use different asns on the same instance (unless
using views).
The result is misleading, because the last asn added in the
configuration is used and it would be hard for users to detect.
Here we enforce the behaviour, rejecting a configuration
that sets different asn for different peers.

Signed-off-by: Federico Paolinelli <fpaoline@redhat.com>

* E2ETests: remove routerid filling

RouterID will make the config be rejected in frr mode, and is a corner
feature. Hence, here we remove it from tests.

Signed-off-by: Federico Paolinelli <fpaoline@redhat.com>