Protect k8s events with a mutex
The core implementation assumes that the events are processed serially,
because originally there was a single queue being fed by all the events.
When switching to controller-runtime we have several controllers, each
one backed up by a different goroutine. To preserve the seriality, we
add a mutex to avoid concurrent processing of k8s events.

Signed-off-by: Federico Paolinelli <fpaoline@redhat.com>