CRDs: add the new layout
This is the evolution of the CRDs defined in the operator, separating
the IP allocation function handled by the address pool, the connection
with bgp peers handled by the BGPPeer CRD, and the intent to advertise a
prefix over L2 or BGP (or both) via the L2Advertisement and
BGPAdvertisement CRDs.

Instead of having a different addersspool version, we create a new CRD
(IPPool) which will live together with the old address pool. This is
required because it's not possible to have converting webhook from one
CRD (addresspool) to two (addresspool, new version + BGP Advertisement
or L2 advertisement). Also, it's not possible to have two different
versions of the same CR stored in etcd. For this reason, it's safer to
have a new CR and make metallb work with the old one.

BGPPeer is then moved to v1beta2, because there it's possible to
translate the node selector, and we'll add new fields in a compatible
manner.

Signed-off-by: Federico Paolinelli <fpaoline@redhat.com>