Remove the avoidbuggyips field
When defining the first version of the CRDs, it was decided not to
expose this flag.
If .0 / .255 ips are problematic, users can define custom ranges to skip
them. Given this is the first version of the api, we are removing the
field in order to propose a slimmer api surface, with the possibility of
readding it later on.

Link to the discussion: https://github.com/metallb/metallb-operator/pull/17#discussion_r652824338

Signed-off-by: Federico Paolinelli <fpaoline@redhat.com>