Tasks: fix the release command
When we splitted the release in release + bumprelease, we changed the
semantic of bumprelease which accepts a string and not a semver object.
This is done to be able to test the bumprelease function without having
to do the whole release process. Here, we fix it to provide the function
the version string coming from outside.

Signed-off-by: Federico Paolinelli <fpaoline@redhat.com>